requirements.txt
asgiref==3.8.1
autopep8==2.3.1
diff-match-patch==20241021
Django==5.1.3
django-admin-interface==0.29.2
django-ckeditor-5==0.2.15
django-colorfield==0.11.0
django-cors-headers==4.6.0
django-filter==24.3
django-import-export==4.3.3
django-js-asset==2.2.0
django-silk==5.3.1
django-taggit==6.1.0
djangorestframework==3.15.2
gprof2dot==2024.6.6
pillow==11.0.0
pycodestyle==2.12.1
python-slugify==8.0.4
sqlparse==0.5.2
tablib==3.7.0
text-unidecode==1.3
tomli==2.2.1
typing_extensions==4.12.2
